Explain how to get updates
Levels that aren't finished have no content. This allows the level to
exist without having to make a dummy commit to include any other
"unfinished" indicators.

A `git fetch` should be sufficient to pull in new banches, and a `git
pull` in the blank level should pull in the new level contents (a fetch
won't update any local branches).

This may be automated in a script or hook in the future.