Add update script
To properly update the game, all game branches must be hard-reset to
their respective upstream branches. A fix in a level branch will often
result in a completely different commit history for the branch and will
need to be replaced.

The output (stdout and stderr) is discarded to avoid spoilers. It is set
using `exec` instead of passing the `-q` flag to `git` for simplicity
and as a catch-all.