Level 12
Code: lemonlime
Next: 2xcompile

False levels: None

Parents:

1. game/lemonlime/master

Purpose:

This level is meant to introduce the player to Git bundles.

Walkthrough:

1. Read the README. Notice the mention of "bundle" and that two branches
   are contained in it. Also note the note at the bottom.
2. Look at the filesystem. Notice the file `2xcompile.bundle`. This is a
   Git bundle.
3. Add the bundle as a remote by running
   `git remote add scratch-level012 2xcompile.bundle`
   (where "scratch-level012" can be almost anything you want as long as
   it starts with `scratch-` as stated in the note in the README).
4. Run `git fetch scratch-level012`.
5. Check out `game/2xcompile/master`.

There are ways to complete this level without creating a remote, but
doing so is the simplest and allows the next level to work as expected.
It is also the usual way of working with bundles.